requirements.txt
pandas
langchain
langchain-openai
langchain-opensearch
opensearch-py

--extra-index-url https://download.pytorch.org/whl/cu130
torch

sentence-transformers
einops

python-telegram-bot